requirements.txt
flask==3.0.0
twilio==8.10.0
openai>=1.12.0
chromadb==0.4.18
python-dotenv==1.0.0
beautifulsoup4==4.12.2
numpy==1.24.3
werkzeug==3.0.1
google-api-python-client==2.100.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
msal==1.24.1
requests==2.32.3
httpx>=0.27.0
python-docx==1.1.0
PyPDF2==3.0.1

